+++ AddOns/BetterBlizzFrames/CHANGELOG.md
@@ -1,3 +1,26 @@
+# BetterBlizzFrames 1.9.3b
+## Midnight
+### Tweak
+- Fix cooldown stuff for upcoming Midnight changes.
+### Bugfix
+- Fix the "Hide Shadow" setting for All Frames not working on rares.
+
+# BetterBlizzFrames 1.9.3
+## Midnight
+### New
+- Add Jazggz profile (www.twitch.tv/jazggz). Thank you for sharing <3
+- New "Improved Target Highlight" setting for PartyFrames (General).
+- New "Hide Dispel Icon" sub-setting for "Hide Dispel Overlay" for PartyFrames (General).
+- New "Change ActionBar CD Size" setting (Misc).
+- New "Hide All Manabar Text" setting (Misc).
+### Tweak
+- Interrupt Icon no longer shows up on un-interruptible casts like how it used to before Midnight.
+- Add few more Edit Mode stuff to the Edit Mode transparancy slider.
+- Change border style a bit on PartyFrames due to a secret issue in Blizzards code.
+### Bugfix
+- Fix secret error issue due to aura borders on partyframe doing some lua errors.
+- Temporary fix with debug print on secret error on party arena names settings. Pls report if you see it.
+
 # BetterBlizzFrames 1.9.2c
 ## Midnight
 ### Tweak